[To dev/1.3] Pipe: Reduced the progress index report interval & Added some logs (#15905) (#15908)
* partial

* Changed default

* Update PipeDataNodeTaskAgent.java

* Next

+++ iotdb-core/node-commons/src/main/java/org/apache/iotdb/commons/pipe/resource/log/PipeLogStatus.java
@@ -29,11 +29,11 @@ class PipeLogStatus {
 
   private final Logger logger;
 
-  private final int maxAverageScale;
+  private final double maxAverageScale;
   private final int maxLogInterval;
   private final AtomicLong currentRounds = new AtomicLong(0);
 
-  PipeLogStatus(final Class<?> logClass, final int maxAverageScale, final int maxLogInterval) {
+  PipeLogStatus(final Class<?> logClass, final double maxAverageScale, final int maxLogInterval) {
     logger = LoggerFactory.getLogger(logClass);
 
     this.maxAverageScale = maxAverageScale;
